==> new_scripts/1_Data_import_exploration.Rmd <==
---
title: "Data Exploration"
author: "Author"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
editor_options:
  chunk_output_type: console
---

# 1. Importing Libraries

```{r libraries, message=FALSE, warning=FALSE}
library(tidyverse)
library(readxl)
library(ggpubr)
library(ggsci)
library(UpSetR)
library(rstatix)
source('functions_cdis_exploration.R')
```

# 2. Import data

```{r set_path, message=FALSE}
#set path variables
project_dir <- getwd()

# Give a name for your project so the results are all grouped together in a directory with the same name
project_name <- 'project'
figures_dir <- file.path(project_dir, paste0(project_name, '_output_figures'))
tables_dir <- file.path(project_dir,  paste0(project_name, '_output_tables'))

# Create output directories
dir.create(figures_dir, showWarnings = FALSE)

dir.create(tables_dir, showWarnings = FALSE)
```


The data for in this script is the **Compounds table** table exported from **Compound Discoverer**. For this script to work export and excel table with the columns: **Name**, **Formula**, **Annot. deltaMass [ppm]**, **Calc. MW**, **RT [min]**, **MS2**, **Area**, **Gap Status** and **Gap Fill Status**.
The **metadata table** can be exported directly from the **Samples tab** in **Compound Discoverer** by copying to a blank text file.


```{r import_data, message=FALSE}
# Import data tables
cd_results_file <- file.path(project_dir, '..', '..', 'data', 'Compound_discoverer.xlsx')
cd_results_table <- read_xlsx(cd_results_file)

# Select columns needed for downstream analysis

cd_results_table <- cd_results_table %>%
  filter(!is.na(Formula)) %>% 
  filter(abs(`Annot. DeltaMass [ppm]`) < 5) %>% 
  arrange(desc(`Calc. MW`)) %>% 
  mutate(FeatureID = paste0('Feature',formatC(n():0001, width = 4, flag = '0'))) %>%
  select(FeatureID, Name, Formula, `Calc. MW`, 
         contains('Annotation source'), contains('Results'), contains('Pathways'), 
         contains('Area:'), contains('Gap Status:'), 
         contains('Labeling Status:'), contains('Gap Fill Status:')) %>% 
  # Differentiate between features that share the same name using "peak#" at the end of the name
  group_by(Name) %>% 
  add_count(Name) %>% 
  # Create variable with names for plotting (useful in following scripts)
  mutate(name4plot = ifelse(is.na(Name), FeatureID, 
                            ifelse(n == 1, Name, paste0(Name, '-peak', n():1)))) %>% 
  select(-n) %>% 
  ungroup()

# Import metadata 
metadata_file <- file.path(project_dir, '..', '..', 'data', 'metadata.csv')
metadata <- read_csv(metadata_file)

# Select only the useful columns and fix column names

metadata <- metadata %>%  
  rename(SampleID = `Sample Identifier`)

table_file <- file.path(tables_dir, 'fixed_metadata.csv')
write_csv(metadata, table_file)
```

# 3. Data Manipulation and thermodynamic indices calculations

```{r data, message=FALSE, warning=FALSE}

# Split formula column into elemental counts
cd_results_table <- separate_formula(cd_results_table)

# Calculate ratios and thermodynamic indices
cd_results_table <- calc_ratios_n_idxs(cd_results_table)

# Calculate classes
cd_results_table <- calc_classes(cd_results_table)

# Gather area under the curve (AUC) values per sample

compounds_table <- cd_results_table %>% 
  select(-contains('Gap Status:'), -contains('Gap Fill Status:')) %>% 
  pivot_longer(contains('Area:'), names_to = 'SampleID', values_to = 'AUC') %>% 
  filter(AUC > 0) %>%  
  mutate(SampleID = str_remove(SampleID, 'Area: '),
         SampleID = str_remove(SampleID, '.raw.*'))

# Save gap-filled table to be used in Statistical Analysis

table_file <- file.path(tables_dir, 'gap_filled_compounds_table.csv')
write_csv(compounds_table, table_file)

```

### Gap filtering 

Removing features that were gap filled is mostly done to see if there are differences in the overall biochemical indexes among samples. All statistical analysis are done with gap-filled data

```{r}

# Gather "Gap Status" for filtering 
  gap_status <- cd_results_table %>% 
    select(FeatureID, contains('Gap Status:')) %>% 
    gather(contains('Gap Status:'), key = 'SampleID', value = 'gap_status')
  gap_status$SampleID <- str_remove(gap_status$SampleID, 'Gap Status: ')
  gap_status$SampleID <- str_remove(gap_status$SampleID, '.raw.*')
  
  ## Filtering
  
  compounds_table <- left_join(compounds_table, gap_status, by = c('FeatureID', 'SampleID')) %>% 
    filter(gap_status != 'Full gap')
  
  # Plotting types of gaps detected
  
  gap_fill_status <- cd_results_table%>% 
    select(FeatureID, Name, contains('Gap Fill Status:')) %>% 
    pivot_longer(contains('Gap Fill Status:'), names_to = 'SampleID', values_to = 'gap_fill') %>% 
    mutate(SampleID = str_remove(SampleID, 'Gap Fill Status: '),
           SampleID = str_remove(SampleID, '.raw.*')) %>% 
    mutate(gap_fill_type = case_when(gap_fill == 32 ~ 'Filled by spectrum noise',
                                     gap_fill == 128 ~ 'Filled by re-detected peak',
                                     gap_fill == 0 ~ 'No gap to fill',
                                     gap_fill == 16 ~ 'Filled by simulated peak',
                                     gap_fill == 8 ~ 'Filled by trace area',
                                     gap_fill == 64 ~ 'Filled by matching ion')) %>% 
    mutate(gap_fill_type = factor(gap_fill_type, levels = c('No gap to fill', 
                                                            'Filled by re-detected peak', 
                                                            'Filled by simulated peak', 
                                                            'Filled by matching ion',
                                                            'Filled by trace area', 
                                                            'Filled by spectrum noise')))

```


``` {r}
# Add metadata information

compounds_table <- left_join(compounds_table, metadata, by = 'SampleID')

table_file <- file.path(tables_dir, 'compounds_table.csv')
write_csv(compounds_table, table_file)

```

# 4. Summary plots

## 4.1. Number of annotated compounds

This figure shows the number of compounds that were detected in each file, and how many of them were assigned a putative structure through database searches

```{r n_identified_masses, message=FALSE}
# Create a new tibble to count the number of compounds per file
summary_table <- compounds_table %>% 
  filter(!is.na(treatment)) %>% 
  select(SampleID, Name) %>% 
  mutate(Status = ifelse(is.na(Name), 'No structure', 'Structure')) %>% 
  group_by(SampleID) %>% 
  count(Status) %>% 
  left_join(metadata, by = c('SampleID')) %>% 
  mutate(Status = factor(Status, levels = c('No structure', 'Structure'))) %>% 
  filter(!is.na(treatment))

id_summary_plot <- plot_col(summary_table, SampleID, n, Status) +
  labs(title = 'Number of masses found',
       x = 'Sample name',
       y = 'Count') +
  theme(plot.title = element_text(face = 'bold', hjust = 0.5),
        axis.text.x = element_text(angle = 45,
                                   hjust = 1)) +
  facet_grid(cols = vars(treatment), scales = 'free_x', space = 'free_x')

id_summary_plot

figure_file <- file.path(figures_dir, 'id_summary.png')
ggsave(figure_file, id_summary_plot, dpi = 300)

```

## 4.3 Gap filled compounds

```{r}

gap_summary_table <- gap_fill_status %>% 
  select(SampleID, gap_fill_type) %>% 
  group_by(SampleID) %>% 
  count(gap_fill_type) %>% 
  left_join(metadata, by = 'SampleID') %>% 
  filter(!is.na(treatment))

gap_summary_plot <- plot_col(gap_summary_table, SampleID, n , gap_fill_type) +
  labs(title = 'Gap filled compounds',
       x = 'Sample name',
       y = 'Count') +
  theme(plot.title = element_text(face = 'bold', hjust = 0.5),
        axis.text.x = element_text(angle = 45,
                                   hjust = 1)) +
  scale_fill_manual(values = c('green4', 'aquamarine3', 'blue2', 
                               'blue', 'purple', 'firebrick2')) +
  facet_grid(cols = vars(treatment), scales = 'free_x', space = 'free_x')

gap_summary_plot

figure_file <- file.path(figures_dir, 'gap_summary.png')
ggsave(figure_file, gap_summary_plot, dpi = 300)

```


## 4.2. Upset plots of number of detected compounds

### Compounds detected in each treatment (without gap filling)

```{r message=FALSE}

# Venn diagram of the compounds present in each time
my_list <- list(Control = unique(compounds_table$FeatureID[compounds_table$treatment == 'CTR']),
                Water_pulse = unique(compounds_table$FeatureID[compounds_table$treatment == 'WP']))

figure_file <- file.path(figures_dir, 'upset_per_treatment.png')
png(figure_file, res = 300, width = 2400, height = 1600)
upset(fromList(my_list), order.by = 'freq')
dev.off()
```

### Compounds detected at each time point (without gap filling)

```{r message=FALSE}

# Venn diagram of the compounds present in each time
my_list <- list(T0 = unique(compounds_table$FeatureID[compounds_table$time == 'T0']),
                T1 = unique(compounds_table$FeatureID[compounds_table$time == 'T1']),
                T2 = unique(compounds_table$FeatureID[compounds_table$time == 'T2']),
                T3 = unique(compounds_table$FeatureID[compounds_table$time == 'T3']),
                T4 = unique(compounds_table$FeatureID[compounds_table$time == 'T4']))

figure_file <- file.path(figures_dir, 'upset_per_time.png')
png(figure_file, res = 300, width = 2400, height = 1600)
upset(fromList(my_list), order.by = 'freq')
dev.off()
```

# 5. Exploratory plots

## 5.1. Van Krevelen Diagrams

### Van Krevelen diagram by treatment


```{r message=FALSE}

# Ven Krevelen Diagram based on time
vank <- compounds_table %>% 
  filter(!is.na(treatment)) %>% 
  plot_vank(., Class, treatment)
vank

figure_file <- file.path(figures_dir, 'Venk_diagram_treatment.jpg')
ggsave(figure_file, vank, dpi = 300)

```

### Van Krevelen diagram by time

```{r message=FALSE}

# Ven Krevelen Diagram based on time
vank <- compounds_table %>% 
  filter(!is.na(treatment)) %>% 
  plot_vank(., Class) +
  facet_wrap(~ time)
vank

figure_file <- file.path(figures_dir, 'Venk_diagram_time.jpg')
ggsave(figure_file, vank, dpi = 300)

```

## 6.1 Gibbs Free Energy (GFE) exploratory plots

### Boxplot of GFE of different treatments

```{r message=FALSE}

stat_table <- compounds_table %>% 
  filter(!is.na(treatment)) %>%
  select(GFE, treatment) %>% 
  wilcox_test(GFE ~ treatment) %>% 
  add_xy_position(step.increase = 3) %>% 
  add_significance()

gfe_boxplot <- compounds_table %>% 
  filter(!is.na(treatment)) %>%
  ggplot(aes(x = treatment,
             y = GFE,
             fill = treatment)) +
  geom_boxplot() +
  scale_fill_brewer(palette = 'Dark2') +
  stat_pvalue_manual(data = stat_table, label = 'p.signif', inherit.aes = FALSE, hide.ns = TRUE) +
  labs(title = 'Gibbs Free Energy (GFE) per habitat') +
  theme_bw() +
  theme(plot.title = element_text(face = 'bold', hjust = 0.5))

gfe_boxplot

figure_file <- file.path(figures_dir, 'gfe_treatment.png')
ggsave(figure_file, gfe_boxplot, dpi = 300)
```

### Boxplot of GFE at different times per treatment

```{r message=FALSE}

stat_table <- compounds_table %>% 
  filter(!is.na(treatment)) %>%
  group_by(treatment) %>% 
  wilcox_test(GFE ~ time) %>% 
  add_xy_position(step.increase = 3) %>% 
  add_significance()

gfe_boxplot <- compounds_table %>% 
  filter(!is.na(treatment)) %>%
  ggplot(aes(x = time,
             y = GFE,
             fill = time)) +
  geom_boxplot() +
  scale_fill_jama() +
  stat_pvalue_manual(data = stat_table, label = 'p.adj.signif', inherit.aes = FALSE, hide.ns = TRUE) +
  labs(title = 'Gibbs Free Energy (GFE) per habitat') +
  theme_bw() +
  theme(plot.title = element_text(face = 'bold', hjust = 0.5)) +
  facet_grid(rows = vars(treatment))

gfe_boxplot

figure_file <- file.path(figures_dir, 'gfe_time_per_treatment.png')
ggsave(figure_file, gfe_boxplot, dpi = 300)
```

### Boxplot of GFE per treatment at different times

```{r message=FALSE}

stat_table <- compounds_table %>% 
  filter(!is.na(treatment),
         time != 'T0') %>%
  group_by(time) %>% 
  wilcox_test(GFE ~ treatment) %>% 
  add_xy_position(step.increase = 3) %>% 
  adjust_pvalue() %>% 
  add_significance()

gfe_boxplot <- compounds_table %>% 
  filter(!is.na(treatment),
         time != 'T0') %>%
  ggplot(aes(x = treatment,
             y = GFE,
             fill = treatment)) +
  geom_boxplot() +
  scale_fill_brewer(palette = 'Dark2') +
  stat_pvalue_manual(data = stat_table, label = 'p.adj.signif', inherit.aes = FALSE, hide.ns = TRUE) +
  labs(title = 'Gibbs Free Energy (GFE) per habitat') +
  theme_bw() +
  theme(plot.title = element_text(face = 'bold', hjust = 0.5)) +
  facet_grid(cols = vars(time))

gfe_boxplot

figure_file <- file.path(figures_dir, 'gfe_time_per_treatment.png')
ggsave(figure_file, gfe_boxplot, dpi = 300)
```
