==> new_scripts/4_Dysregulated_Pathways.Rmd <==
---
title: "RP Lab - Dysregulated Pathways"
author: "Author"
date: "2/24/2022"
output: html_document
---

This Notebook will explore the KEGG (and Metabolika) Pathways that are dysregulated

# 1. Importing Libraries

```{r libraries, message=FALSE, warning=FALSE}
library(ggpubr)
library(ggsci)
library(ggrepel)
library(pheatmap)
library(RColorBrewer)
library(scales)
library(plyr)
library(cowplot)
library(readxl)
library(KEGGREST)
library(tidyverse)
source('functions_dysreg_path.R')
```

# 2. Import Data

```{r set_path, message=FALSE, warning=FALSE}
# set path variables
project_dir <- getwd()
project_name <- 'project'
figures_dir <- file.path(project_dir, paste0(project_name, '_output_figures'))
tables_dir <- file.path(project_dir,  paste0(project_name, '_output_tables')) 
```

This analysis uses the tables generated during the Differential Analysis script.

```{r message=FALSE, warning=FALSE}
# Load compounds table

compounds_table_file <- file.path(tables_dir, 'gap_filled_compounds_table.csv')

compounds_table <- read_csv(compounds_table_file)

# Import metadata and fix names
metadata_file <- file.path(tables_dir, 'fixed_metadata.csv')
metadata <- read_csv(metadata_file)

# Import table(s) obtained during the Differential analysis
table_file <- file.path(tables_dir, 'de.ctr.t0_t1.csv')
de.ctr.t0_t1 <- read_csv(table_file)

table_file <- file.path(tables_dir, 'de.ctr.t0_t2.csv')
de.ctr.t0_t2 <- read_csv(table_file)

table_file <- file.path(tables_dir, 'de.ctr.t0_t3.csv')
de.ctr.t0_t3 <- read_csv(table_file)

table_file <- file.path(tables_dir, 'de.ctr.t0_t4.csv')
de.ctr.t0_t4 <- read_csv(table_file)


# Import table(s) obtained during the Differential analysis
table_file <- file.path(tables_dir, 'de.wp.t0_t1.csv')
de.wp.t0_t1 <- read_csv(table_file)

table_file <- file.path(tables_dir, 'de.wp.t0_t2.csv')
de.wp.t0_t2 <- read_csv(table_file)

table_file <- file.path(tables_dir, 'de.wp.t0_t3.csv')
de.wp.t0_t3 <- read_csv(table_file)

table_file <- file.path(tables_dir, 'de.wp.t0_t4.csv')
de.wp.t0_t4 <- read_csv(table_file)



```

From Compound Discover export a **2-level** table with: 

- *KEGG Pathways* in the first level with the columns: PathwayID, Pathway Name
- *Compounds* in the second level (Only way I found to be able to export the **KEGG Compounds IDs**) with the columns: Checked, Calc. MW, Name, Formula, KEGG Compound Ids, KEGG Compound Names

```{r message=FALSE, warning=FALSE}
kegg_table_file <- file.path('..', '..', 'data', 'kegg_annot.xlsx')
kegg_table <- read_xlsx(kegg_table_file)

# Fix column names
colnames(kegg_table) <- c('PathwayID', 'Pathway_Name', 'Name', 'Formula', 'deltamass', 'Calc. MW', 'RT', 'KEGG_ID', 'KEGG_Name')

# Fix 2-level table into regular table

kegg_table <- kegg_table %>% 
  select(-RT) %>% 
  filter(Pathway_Name != 'Checked') %>% 
  mutate(Pathway_Name = ifelse(Pathway_Name == 'FALSE', NA, Pathway_Name)) %>% 
  fill(PathwayID, Pathway_Name, .direction = 'down') %>% 
  filter(!is.na(`Calc. MW`)) %>%  
  mutate(deltamass = as.numeric(deltamass)) %>% 
  filter(abs(deltamass) < 5) %>% 
  select(-`Calc. MW`, - deltamass)

comp_kegg <- left_join(compounds_table, kegg_table, by = c('Name', 'Formula')) %>% 
  filter(!(str_detect(SampleID, 'Blank') | str_detect(SampleID, 'Pooled'))) %>% 
  distinct()

table_file <- file.path(tables_dir, 'Compounds_with_KEGG.csv')
write_csv(comp_kegg, table_file)
```

# 3. Analyze KEGG Pathways annotated at each time point

Merge KEGG annotation with Differential analysis data

```{r}
# This step includes filtering for significant DE metabolites (pval.adj < 0.05)
kegg.ctr.t0_t1 <- de.ctr.t0_t1 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'CTR - T1 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.ctr.t0_t1.csv')
write_csv(kegg.ctr.t0_t1, table_file)

kegg.ctr.t0_t2 <- de.ctr.t0_t2 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'CTR - t2 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.ctr.t0_t2.csv')
write_csv(kegg.ctr.t0_t2, table_file)


kegg.ctr.t0_t3 <- de.ctr.t0_t3 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'CTR - t3 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.ctr.t0_t3.csv')
write_csv(kegg.ctr.t0_t3, table_file)

kegg.ctr.t0_t4 <- de.ctr.t0_t4 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'CTR - t4 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.ctr.t0_t4.csv')
write_csv(kegg.ctr.t0_t4, table_file)


kegg.wp.t0_t1 <- de.wp.t0_t1 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'wp - T1 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.wp.t0_t1.csv')
write_csv(kegg.wp.t0_t1, table_file)

kegg.wp.t0_t2 <- de.wp.t0_t2 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'wp - t2 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.wp.t0_t2.csv')
write_csv(kegg.wp.t0_t2, table_file)


kegg.wp.t0_t3 <- de.wp.t0_t3 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'wp - t3 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.wp.t0_t3.csv')
write_csv(kegg.wp.t0_t3, table_file)

kegg.wp.t0_t4 <- de.wp.t0_t4 %>%
  select(FeatureID, Name, Formula, log2FC, pval.adj, Comment) %>% 
  filter(abs(log2FC) > 1, pval.adj < 0.1) %>% 
  left_join(kegg_table, by = c('Name', 'Formula')) %>% 
  distinct() %>% 
  filter(!is.na(Pathway_Name)) %>% 
  mutate(comparison = 'wp - t4 vs T0')

table_file <- file.path(tables_dir, 'kegg_table.wp.t0_t4.csv')
write_csv(kegg.wp.t0_t4, table_file)



```

## 3.1 Plot number of KEGG pathways

```{r}
#Acceptable KEGG pathways
table_file <- file.path('..', '..', 'data', 'all_paths.csv')
pathways <- read_csv(table_file)

# Set colors for plot
my_colors = c('Upregulated' = '#d34849',
              'Downregulated' = '#005193')

kegg.counts.ctr.t0_t1 <- kegg.ctr.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t1')

kegg.plot.ctr.t0_t1 <- kegg.counts.ctr.t0_t1 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'CTR - t0 vs t1',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.ctr.t0_t1


figure_file <- file.path(figures_dir, 'diff_KEGG.ctr.t0_t1.png')
ggsave(figure_file, kegg.plot.ctr.t0_t1, dpi = 300, height = 7)

kegg.counts.ctr.t0_t2 <- kegg.ctr.t0_t2 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t2')

kegg.counts.ctr.t0_t1 <- kegg.ctr.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t1')

kegg.plot.ctr.t0_t2 <- kegg.counts.ctr.t0_t2 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'CTR - t0 vs t2',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.ctr.t0_t2


figure_file <- file.path(figures_dir, 'diff_KEGG.ctr.t0_t2.png')
ggsave(figure_file, kegg.plot.ctr.t0_t2, dpi = 300, height = 7)

kegg.counts.ctr.t0_t3 <- kegg.ctr.t0_t3 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t3')

kegg.counts.ctr.t0_t1 <- kegg.ctr.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t1')

kegg.plot.ctr.t0_t3 <- kegg.counts.ctr.t0_t3 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'CTR - t0 vs t3',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.ctr.t0_t3


figure_file <- file.path(figures_dir, 'diff_KEGG.ctr.t0_t3.png')
ggsave(figure_file, kegg.plot.ctr.t0_t3, dpi = 300, height = 7)

kegg.counts.ctr.t0_t4 <- kegg.ctr.t0_t4 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t4')

kegg.counts.ctr.t0_t1 <- kegg.ctr.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'CTR - t0 vs t1')

kegg.plot.ctr.t0_t4 <- kegg.counts.ctr.t0_t4 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'CTR - t0 vs t4',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.ctr.t0_t4


figure_file <- file.path(figures_dir, 'diff_KEGG.ctr.t0_t4.png')
ggsave(figure_file, kegg.plot.ctr.t0_t4, dpi = 300, height = 7)

```

### WP

```{r}

#Acceptable KEGG pathways
table_file <- file.path('..', '..', 'data', 'all_paths.csv')
pathways <- read_csv(table_file)

# Set colors for plot
my_colors = c('Upregulated' = '#d34849',
              'Downregulated' = '#005193')

kegg.counts.wp.t0_t1 <- kegg.wp.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t1')

kegg.plot.wp.t0_t1 <- kegg.counts.wp.t0_t1 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'wp - t0 vs t1',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.wp.t0_t1


figure_file <- file.path(figures_dir, 'diff_KEGG.wp.t0_t1.png')
ggsave(figure_file, kegg.plot.wp.t0_t1, dpi = 300, height = 7)

kegg.counts.wp.t0_t2 <- kegg.wp.t0_t2 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t2')

kegg.counts.wp.t0_t1 <- kegg.wp.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t1')

kegg.plot.wp.t0_t2 <- kegg.counts.wp.t0_t2 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'wp - t0 vs t2',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.wp.t0_t2


figure_file <- file.path(figures_dir, 'diff_KEGG.wp.t0_t2.png')
ggsave(figure_file, kegg.plot.wp.t0_t2, dpi = 300, height = 7)

kegg.counts.wp.t0_t3 <- kegg.wp.t0_t3 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t3')

kegg.counts.wp.t0_t1 <- kegg.wp.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t1')

kegg.plot.wp.t0_t3 <- kegg.counts.wp.t0_t3 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'wp - t0 vs t3',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.wp.t0_t3


figure_file <- file.path(figures_dir, 'diff_KEGG.wp.t0_t3.png')
ggsave(figure_file, kegg.plot.wp.t0_t3, dpi = 300, height = 7)

kegg.counts.wp.t0_t4 <- kegg.wp.t0_t4 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t4')

kegg.counts.wp.t0_t1 <- kegg.wp.t0_t1 %>% 
  group_by(Comment) %>% 
  count(Pathway_Name) %>% 
  mutate(axis = ifelse(Comment == 'Upregulated', n, n*-1),
         comparison = 'wp - t0 vs t1')

kegg.plot.wp.t0_t4 <- kegg.counts.wp.t0_t4 %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways',
       subtitle = 'wp - t0 vs t4',
       y = 'Pathway Name') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5)) +
  scale_x_continuous(limits = c(-15, 15))
kegg.plot.wp.t0_t4


figure_file <- file.path(figures_dir, 'diff_KEGG.wp.t0_t4.png')
ggsave(figure_file, kegg.plot.wp.t0_t4, dpi = 300, height = 7)
```

### 3.1.3 Combined graph

```{r}

kegg.counts.combined <- rbind(kegg.counts.ctr.t0_t1, kegg.counts.ctr.t0_t2, kegg.counts.ctr.t0_t3, kegg.counts.ctr.t0_t4,
                              kegg.counts.wp.t0_t1, kegg.counts.wp.t0_t2, kegg.counts.wp.t0_t3, kegg.counts.wp.t0_t4)

kegg.counts.combined[428,] <- list(NA, 'Metabolic pathways', NA, NA, 'CTR - t0 vs t1')

selected_paths <- kegg.counts.combined %>% 
  filter(abs(axis) > 2) %>% 
  pull(Pathway_Name)

kegg.plot.combined <- kegg.counts.combined %>% 
  filter(Pathway_Name %in% pathways$KEGG_pathway,
         Pathway_Name %in% selected_paths) %>% 
  ggplot() +
  geom_col(aes(x = axis,
               y = fct_reorder(Pathway_Name, n),
               fill = Comment),
           color = 'black') +
  geom_text(aes(label = n,
                x = axis/2,
                y = fct_reorder(Pathway_Name, n)),
            color = 'white',
            fontface = 'bold',
            size = 3) +
  scale_fill_manual(values = my_colors) +
  labs(title = '# of Metabolites in KEGG Pathways') +
  theme_bw() +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        plot.title = element_text(face = 'bold', hjust = 0.5),
        legend.position = 'none') +
  facet_wrap(~comparison, nrow = 2)
kegg.plot.combined

figure_file <- file.path(figures_dir, 'diff_KEGG.combined.png')
ggsave(figure_file, kegg.plot.combined, dpi = 300, height = 7, width = 12)

```

## 3.2 Creating tables for KEGG ID mapper

```{r}
kegg_dir <- file.path(tables_dir, 'kegg_mapper')
dir.create(kegg_dir)

paths.during_pre <- kegg.counts.during_pre %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  pull(Pathway_Name)

for(path in unique(paths.during_pre)){
  df <- kegg.during_pre %>% 
    filter(Pathway_Name == path) %>% 
    select(FeatureID, log2FC, KEGG_ID) %>% 
    separate_rows(KEGG_ID, sep = '; ') %>% 
    mutate(color = ifelse(log2FC > 0, 'red', 'blue'),
           enter_this = paste(KEGG_ID, color, sep = ' ')) %>% 
    select(KEGG_ID, enter_this) %>% 
    distinct()
  
  table_file <- file.path(kegg_dir, paste0('During_vs_pre_', path, 'ids_for_mapper.csv'))
  
  write_csv(df, table_file)
  print(paste(path, 'done'))
}


paths.post_pre <- kegg.counts.post_pre %>% 
  filter(n > 0,
         Pathway_Name %in% pathways$KEGG_pathway) %>% 
  pull(Pathway_Name)

for(path in unique(paths.post_pre)){
  df <- kegg.post_pre %>% 
    filter(Pathway_Name == path) %>% 
    select(FeatureID, log2FC, KEGG_ID) %>% 
    separate_rows(KEGG_ID, sep = '; ') %>% 
    mutate(color = ifelse(log2FC > 0, 'red', 'blue'),
           enter_this = paste(KEGG_ID, color, sep = ' ')) %>% 
    select(KEGG_ID, enter_this) %>% 
    distinct()
  
  table_file <- file.path(kegg_dir, paste0('Post_vs_pre_', path, 'ids_for_mapper.csv'))
  
  write_csv(df, table_file)
  print(paste(path, 'done'))
}

```
